wireless sensor network
A light-weight routing protocol for an Ad Hoc wireless sensor network
based on Contiki environment. There are 4 nodes exists in this network.
Two for data transmission (sender and receiver), the others for routing
which build a bridge between sender and receiver. The programming
language used for the program is C, due to Contiki environment. The
program implements several specified requirements, all of which were
implemented and tested. Router, sender and receiver could successfully
satisfiy all requirements. However, there are some issues occured during
the configuration of development environment as well as the
implementaton process.